Add the ImageJ.png from the 32-bit Linux archive
While at it, also delete a Mac-specific file that slipped into the
Windows installer.

We leave out the 'run' and 'script' shell scripts because we want
to use the unified ImageJ launcher instead.

We also prefer the bezier_tool.jar from the Windows installer
because manual inspection reveals that it seems to be newer (with
ugly commented-out code being superseded by other code).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>