Add links after building
This needs a Groovy Postbuild script:

	// determine whether any axis besides the current one is still building
	isBuilding = false
	manager.build.getParentBuild().getRuns().each() {
		run ->
		if (!run.equals(manager.build) && run.isBuilding()) {
			isBuilding = true
		}
	}

	// get the path to the postbuild script
	root = manager.build.getParentBuild().getProject().getBuildDir()
	postBuild = new hudson.FilePath(root).child("../workspace/postbuild.groovy")

	// run the postbuild script only after all the artifacts were made
	if (!isBuilding && postBuild != null && postBuild.exists()) {
		shell = new GroovyShell(new Binding(manager: manager))
		shell.evaluate(new java.io.InputStreamReader(postBuild.read()))
	}

The reason is that we will only know *after* the build steps whether the
build has been successful. That's the job of the postbuild script.

We need to take care, however, to run this only when all the axes have
been built. Hence we skip the postbuild until all axes are done.

Also, since the postbuild script is run on the main Jenkins node, we need
to make sure to access the correct node's workspace to find the script
(unfortunately, MatrixRun#getWorkspace() does not return the workspace of
the node associated with that particular MatrixRun, hence we have to use
knowledge that the 'master' node actually runs on the same machine).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>